+++ src/app/Balloon.App.Notification/NodeMessage.php
@@ -56,7 +56,10 @@ class NodeMessage implements MessageInterface
      */
     public function getSubject(User $user): string
     {
-        return $this->template->parseSubjectTemplate($this->type, $user, $this->node);
+        return $this->template->getSubject($this->type, [
+            'user' => $user,
+            'node' => $this->node,
+        ]);
     }
 
     /**
@@ -64,11 +67,20 @@ class NodeMessage implements MessageInterface
      */
     public function getBody(User $user): string
     {
-        return $this->template->parseBodyTemplate($this->type, $user, $this->node);
+        return $this->template->getBody($this->type, [
+            'user' => $user,
+            'node' => $this->node,
+        ]);
     }
 
-    public function getMailBody(User $user): string
+    /**
+     * {@inheritdoc}
+     */
+    public function renderTemplate(string $template, User $user): string
     {
-        return $this->template->parseMailBodyTemplate($this->type, $user, $this->node);
+        return $this->template->renderTemplate($this->type, $template, [
+            'user' => $user,
+            'node' => $this->node,
+        ]);
     }
 }