+++ src/app/Balloon.App.Sharelink/Api/v2/ShareLink.php
@@ -65,9 +65,8 @@ class ShareLink extends Controller
      * @apiUse _getNode
      * @apiUse _writeAction
      *
-     * @apiParam (POST Parameter) {object} [options] Sharing options
-     * @apiParam (POST Parameter) {number} [options.expiration] Expiration unix timestamp of the sharing link
-     * @apiParam (POST Parameter) {string} [options.password] Protected shared link with password
+     * @apiParam (POST Parameter) {number} [expiration] Expiration unix timestamp of the sharing link
+     * @apiParam (POST Parameter) {string} [password] Protected shared link with password
      *
      * @apiExample (cURL) example:
      * curl -XPOST "https://SERVER/api/v2/node/share-link?id=544627ed3c58891f058b4686&pretty"
@@ -86,12 +85,10 @@ class ShareLink extends Controller
      *
      * @return Response
      */
-    public function post(?string $id = null, ?string $p = null, array $options = []): Response
+    public function post(?string $id = null, ?string $p = null, ?string $password = null, ?string $expiration = null): Response
     {
         $node = $this->fs->getNode($id, $p);
-        $options['shared'] = true;
-
-        $this->sharelink->shareLink($node, $options);
+        $this->sharelink->shareLink($node, $expiration, $password);
         $result = $this->node_decorator->decorate($node);
 
         return (new Response())->setCode(200)->setBody($result);
@@ -121,9 +118,7 @@ class ShareLink extends Controller
     public function delete(?string $id = null, ?string $p = null): Response
     {
         $node = $this->fs->getNode($id, $p);
-        $options = ['shared' => false];
-
-        $this->sharelink->shareLink($node, $options);
+        $this->sharelink->deleteShareLink($node);
 
         return (new Response())->setCode(204);
     }